refactor(sedona): move url_table resolver into the sedona crate
The URL-as-table resolver is a catalog concern whose only production caller (context.rs) already lives in the sedona crate, so move url_table.rs there from sedona-datasource, repoint its imports to sedona_datasource::{format, provider}, and move the url-table tests alongside it. datafusion-session moves to the sedona crate; sedona-datasource no longer depends on it.
