+++ rust/sedona-datasource/Cargo.toml
@@ -47,7 +47,6 @@ datafusion-common = { workspace = true }
 datafusion-datasource = { workspace = true }
 datafusion-execution = { workspace = true }
 datafusion-physical-expr = { workspace = true }
-datafusion-session = { workspace = true }
 datafusion-physical-plan = { workspace = true }
 futures = { workspace = true }
 object_store = { workspace = true }
