+++ rust/sedona-datasource/src/lib.rs
@@ -18,5 +18,4 @@
 pub mod format;
 pub mod provider;
 pub mod spec;
-pub mod url_table;
 pub mod utility;
