+++ rust/sedona/Cargo.toml
@@ -62,10 +62,12 @@ aws-config = { version = "1.5.17", optional = true }
 aws-credential-types = { version = "1.2.0", optional = true }
 comfy-table = { workspace = true }
 datafusion = { workspace = true, default_features = false, features = ["sql", "parquet", "nested_expressions"] }
+datafusion-catalog = { workspace = true }
 datafusion-common = { workspace = true }
 datafusion-expr = { workspace = true }
 datafusion-ffi = { workspace = true }
 datafusion-optimizer = { workspace = true }
+datafusion-session = { workspace = true }
 dirs = { workspace = true }
 futures = { workspace = true }
 geo-traits = { workspace = true }
